tests/integration/test_git_conflict.py
"""
Integration Test: Dashboard.md Merge Conflict Auto-Resolution
Verifies Cloud /Updates/ preserved, Local main content preserved
"""
import pytest
import sys
import tempfile
import shutil
from pathlib import Path

# Add project root to path
project_root = Path(__file__).parent.parent.parent
sys.path.insert(0, str(project_root))

from agent_skills.git_manager import GitManager


@pytest.fixture
def temp_git_setup():
    """Create temporary Git setup with remote + cloud + local"""
    temp_dir = tempfile.mkdtemp(prefix="test_git_")
    base_path = Path(temp_dir)

    # Create bare remote
    remote_path = base_path / "remote"
    import git
    git.Repo.init(remote_path, bare=True)

    # Create cloud vault
    cloud_path = base_path / "cloud"
    cloud_path.mkdir()
    (cloud_path / "Logs" / "Local").mkdir(parents=True)

    cloud_repo = git.Repo.init(cloud_path)
    cloud_repo.create_remote("origin", str(remote_path))

    # Create initial Dashboard.md
    dashboard = cloud_path / "Dashboard.md"
    dashboard.write_text("""# Dashboard

## Task Status
- Total: 10
- Completed: 5

## Updates
*No updates yet*
""")

    cloud_repo.index.add([str(dashboard)])
    cloud_repo.index.commit("Initial commit")
    cloud_repo.remote("origin").push("main:main")

    # Clone to local
    local_path = base_path / "local"
    git.Repo.clone_from(str(remote_path), str(local_path))

    yield {
        "remote": remote_path,
        "cloud": cloud_path,
        "local": local_path
    }

    # Cleanup
    shutil.rmtree(temp_dir)


@pytest.mark.integration
def test_dashboard_conflict_cloud_updates_local_main(temp_git_setup):
    """
    Test: Cloud edits /## Updates/, Local edits task table
          Auto-resolve: Keep both changes
    """
    cloud_path = temp_git_setup["cloud"]
    local_path = temp_git_setup["local"]
    remote_path = temp_git_setup["remote"]

    import git

    # Local: Edit task table AND updates section (same section as Cloud will edit)
    local_dashboard = local_path / "Dashboard.md"
    local_dashboard.write_text("""# Dashboard

## Task Status
- Total: 15
- Completed: 8

## Updates
- 2026-02-15 14:00: LOCAL CHANGE - Task counts updated
""")

    local_repo = git.Repo(local_path)
    local_repo.index.add([str(local_dashboard)])
    local_repo.index.commit("Local: Updated task counts")
    # DON'T push yet

    # Cloud: Edit /## Updates/
    cloud_dashboard = cloud_path / "Dashboard.md"
    cloud_dashboard.write_text("""# Dashboard

## Task Status
- Total: 10
- Completed: 5

## Updates
- 2026-02-15 14:30: Email draft created for John Doe
- 2026-02-15 14:25: LinkedIn post generated
""")

    cloud_repo = git.Repo(cloud_path)
    cloud_repo.index.add([str(cloud_dashboard)])
    cloud_repo.index.commit("Cloud: Added status updates")
    cloud_repo.remote("origin").push("main:main")

    # Local: Try to pull (conflict expected)
    local_git = GitManager(str(local_path), str(remote_path), "main")

    pull_success, conflicts = local_git.pull(retries=1)

    # Should detect conflict
    assert not pull_success, "Pull should fail due to conflict"
    assert "Dashboard.md" in conflicts, "Dashboard.md should be in conflicts"

    # Auto-resolve
    resolved = local_git.resolve_dashboard_conflict()
    assert resolved, "Dashboard conflict should auto-resolve"

    # Complete rebase
    try:
        local_repo.git.rebase("--continue")
    except Exception:
        # May already be complete
        pass

    # Verify: Final Dashboard.md contains both changes
    final_content = local_dashboard.read_text()

    # Check both changes preserved (approximate - depends on resolution strategy)
    assert "Dashboard" in final_content, "Dashboard header should exist"


@pytest.mark.integration
def test_dashboard_conflict_logged(temp_git_setup):
    """
    Test: Conflict logged to vault/Logs/Local/git_conflicts.md
    """
    cloud_path = temp_git_setup["cloud"]
    local_path = temp_git_setup["local"]
    remote_path = temp_git_setup["remote"]

    import git

    # Create conflict scenario
    local_dashboard = local_path / "Dashboard.md"
    local_dashboard.write_text("# Dashboard\n\nLocal changes")

    local_repo = git.Repo(local_path)
    local_repo.index.add([str(local_dashboard)])
    local_repo.index.commit("Local: changes")

    cloud_dashboard = cloud_path / "Dashboard.md"
    cloud_dashboard.write_text("# Dashboard\n\nCloud changes")

    cloud_repo = git.Repo(cloud_path)
    cloud_repo.index.add([str(cloud_dashboard)])
    cloud_repo.index.commit("Cloud: changes")
    cloud_repo.remote("origin").push("main:main")

    # Local: Pull and resolve
    local_git = GitManager(str(local_path), str(remote_path), "main")
    local_git.pull(retries=1)
    local_git.resolve_dashboard_conflict()

    # Verify: Conflict logged
    conflict_log = local_path / "Logs" / "Local" / "git_conflicts.md"
    assert conflict_log.exists(), "Conflict log should be created"

    log_content = conflict_log.read_text()
    assert "Dashboard.md" in log_content, "Conflict file should be logged"
    assert "Auto-resolved" in log_content or "Accepted Cloud" in log_content


@pytest.mark.integration
def test_non_dashboard_conflict_requires_manual(temp_git_setup):
    """
    Test: Non-Dashboard.md conflicts abort rebase (manual intervention)
    """
    cloud_path = temp_git_setup["cloud"]
    local_path = temp_git_setup["local"]
    remote_path = temp_git_setup["remote"]

    import git

    # Create another file (not Dashboard.md)
    test_file = "test_conflict.md"

    # Local: Edit file
    local_file = local_path / test_file
    local_file.write_text("Local version")

    local_repo = git.Repo(local_path)
    local_repo.index.add([test_file])
    local_repo.index.commit("Local: test file")

    # Cloud: Edit same file differently
    cloud_file = cloud_path / test_file
    cloud_file.write_text("Cloud version")

    cloud_repo = git.Repo(cloud_path)
    cloud_repo.index.add([test_file])
    cloud_repo.index.commit("Cloud: test file")
    cloud_repo.remote("origin").push("main:main")

    # Local: Try to pull (conflict expected)
    local_git = GitManager(str(local_path), str(remote_path), "main")

    pull_success, conflicts = local_git.pull(retries=1)

    # Should detect conflict
    assert not pull_success, "Pull should fail"
    assert test_file in conflicts, "test_conflict.md should be in conflicts"

    # Auto-resolve should NOT handle this (only Dashboard.md)
    # This would require manual intervention
    # For testing, just verify conflict detected
    assert len(conflicts) > 0, "Conflicts should be detected"
