requirements.txt
# Platinum Tier Dependencies
# Inherits ALL Gold tier dependencies + adds dual-agent capabilities

# ====== INHERITED FROM GOLD (and Silver, Bronze) ======
# Core AI
anthropic>=0.18.0                    # Claude API for AI analysis and drafts

# Browser automation
playwright>=1.40.0                   # WhatsApp Web, Gmail automation

# File system and monitoring
watchdog>=3.0.0                      # Vault file monitoring
filelock>=3.12.0                     # Vault write conflict resolution

# Data formats
pyyaml>=6.0                          # YAML frontmatter parsing

# Google APIs
google-api-python-client>=2.80.0    # Gmail API for email monitoring

# Async and HTTP
aiohttp>=3.9.0                       # Async HTTP client
requests>=2.31.0                     # Synchronous HTTP (LinkedIn API, etc.)

# Environment and scheduling
python-dotenv>=1.0.0                 # Environment variable management
schedule>=1.2.0                      # CEO Briefing scheduling

# ====== PLATINUM TIER ADDITIONS ======

# Git synchronization (CRITICAL for Platinum)
gitpython>=3.1.40                    # Git automation for vault sync

# Security and authentication
bcrypt>=4.1.2                        # Dashboard password hashing

# Process management (Python wrapper for PM2 monitoring)
psutil>=5.9.6                        # Process monitoring (CPU, memory for health checks)

# Testing
pytest>=7.4.0                        # Unit and integration testing
pytest-asyncio>=0.21.0               # Async test support

# Note: Next.js dependencies managed separately in nextjs_dashboard/package.json
# Note: Playwright browsers must be installed separately: playwright install chromium
# Note: PM2 installed globally via npm: npm install -g pm2
